chore: move utils.js into frontend/utils.js and update imports
